+++ bsctl/util/k8s/io_streams.go
@@ -1,7 +1,6 @@
 package k8s
 
 import (
-	"bytes"
 	"io"
 
 	genericCLIOptions "k8s.io/cli-runtime/pkg/genericclioptions"
@@ -42,65 +41,5 @@ func IOStreamsFromK8s(options *genericCLIOptions.IOStreams) IOStreams {
 		InObj:     options.In,
 		OutObj:    options.Out,
 		ErrOutObj: options.ErrOut,
-	}
-}
-
-// StreamStores - input/output stream stores
-type StreamStores struct {
-	In     *bytes.Buffer
-	Out    *bytes.Buffer
-	ErrOut *bytes.Buffer
-}
-
-// NewStreamStores - create new stream stores as buffers
-func NewStreamStores() *StreamStores {
-	return &StreamStores{
-		In:     bytes.NewBuffer(nil),
-		Out:    bytes.NewBuffer(nil),
-		ErrOut: bytes.NewBuffer(nil),
-	}
-}
-
-// IOStreamsStore - store for IOStreams
-type IOStreamsStore struct {
-	IOStreams    IOStreams
-	StreamStores *StreamStores
-}
-
-// GetIOStreams - get IOStreams
-func (s *IOStreamsStore) GetIOStreams() IOStreams {
-	return s.IOStreams
-}
-
-// GetStreamStores - get StreamStores
-func (s *IOStreamsStore) GetStreamStores() *StreamStores {
-	return s.StreamStores
-}
-
-// StreamStoresWrapper - lets StreamStores be used as IOStreams
-type StreamStoresWrapper struct {
-	StreamStores *StreamStores
-}
-
-// In - get input stream
-func (s *StreamStoresWrapper) In() io.Reader {
-	return s.StreamStores.In
-}
-
-// Out - get output stream
-func (s *StreamStoresWrapper) Out() io.Writer {
-	return s.StreamStores.Out
-}
-
-// ErrOut - get error output stream
-func (s *StreamStoresWrapper) ErrOut() io.Writer {
-	return s.StreamStores.ErrOut
-}
-
-// NewStreamsStoreWrapper - create a new StreamStoresWrapper
-// to use StreamStores as IOStreams
-func NewStreamsStoreWrapper(stores *StreamStores) IOStreams {
-	return &StreamStoresWrapper{
-		StreamStores: stores,
 	}
 }
