+++ bsctl/util/k8s/io_streams_getter_test.go
@@ -29,7 +29,7 @@ func TestGetIOStreamsGetterConfigured(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			// Arrange
-			expected := &IOStreamsStore{}
+			expected := &IOStreamsAndStores{}
 			realStreams := &IOStreamsConcrete{
 				InObj:     os.Stdin,
 				OutObj:    os.Stdout,
@@ -58,7 +58,7 @@ func TestGetIOStreamsGetterConfigured(t *testing.T) {
 func TestGetStoreOnlyStreams(t *testing.T) {
 	// Arrange
 	stores := NewStreamStores()
-	expected := &IOStreamsStore{
+	expected := &IOStreamsAndStores{
 		IOStreams:    NewStreamsStoreWrapper(stores),
 		StreamStores: stores,
 	}
